Title: 최솟값 3개, Time: 1292ms, Memory: 33MB, Status: Passed - CodeTree

+++ 241001/최솟값 3개/top-3-smallest-number.java
@@ -35,7 +35,7 @@ public class Main {
                 threeNums[i] = pq.poll();
             }
 
-            int multiplyResult = 1;
+            long multiplyResult = 1;
             // 다시 넣기
             for (int i = 0; i < 3; i++) {
                 multiplyResult *= threeNums[i];
